Change RSSI type to Int8

+++ Sonar/Contact Events/ContactEvent.swift
@@ -11,11 +11,11 @@ struct ContactEvent: Equatable, Codable {
 
     var encryptedRemoteContactId: Data? = nil // TODO: This isn't that now, it's a "broadcast payload" and the encrypted id is inside it
     private (set) var timestamp: Date = Date()
-    private (set) var rssiValues: [Int] = []
+    private (set) var rssiValues: [Int8] = []
     private (set) var rssiIntervals: [TimeInterval] = []
     private (set) var duration: TimeInterval = 0
     
-    mutating func recordRSSI(_ rssi: Int, timestamp: Date = Date()) {
+    mutating func recordRSSI(_ rssi: Int8, timestamp: Date = Date()) {
         rssiValues.append(rssi)
         rssiIntervals.append(timestamp.timeIntervalSince(self.timestamp.addingTimeInterval(duration)))
         duration = timestamp.timeIntervalSince(self.timestamp)
